Refactored part one & two
