+++ Part 1/client/src/components/card/Card.scss
@@ -3,6 +3,21 @@
   height: 550px;
   border: 15px solid white;
   position: relative;
+  width: 25%;
+
+  flex-direction: column;
+  align-items: flex-end;
+
+  .card__img-container {
+    height: 300px;
+    width: 500px;
+    margin: 0 auto;
+    width: 100%;
+    img {
+      height: 100%;
+      width: 100%;
+    }
+  }
 
   &:hover {
     .card__cta {
@@ -57,71 +72,63 @@
       rgba(255, 255, 255, 1) 100%
     );
   }
+
+  &:first-child {
+    width: 50%;
+  }
+
+  &:nth-child(7n) {
+    width: 50%;
+  }
+
+  &:nth-child(6n) {
+    width: 50%;
+    flex-direction: row;
+    align-items: flex-start;
+
+    .header-primary {
+      padding-top: 0rem;
+    }
+    .card__img-container {
+      height: 100%;
+      width: 50%;
+      margin: 0 auto;
+      img {
+        height: 100%;
+        width: 100%;
+      }
+    }
+
+    .card__text-box {
+      width: 50%;
+      height: 100%;
+      padding-left: 2rem;
+    }
+  }
 }
 
 /* Styles for Horizontal and vertical cards*/
 
-.horizontal {
-  flex-direction: row;
-  align-items: flex-start;
-
-  .header-primary {
-    padding-top: 0rem;
-  }
-  .card__img-container {
-    height: 100%;
-    width: 50%;
-    margin: 0 auto;
-    img {
-      height: 100%;
-      width: 100%;
-    }
-  }
-
-  .card__text-box {
-    width: 50%;
-    height: 100%;
-    padding-left: 2rem;
-  }
-}
-
-.vertical {
-  flex-direction: column;
-  align-items: flex-end;
-
-  .card__img-container {
-    height: 300px;
-    width: 500px;
-    margin: 0 auto;
-    width: 100%;
-    img {
-      height: 100%;
-      width: 100%;
-    }
-  }
-}
-
 @mixin card-med() {
   .card {
-    width: 0px !important;
-    flex-grow: 1;
     border: 10px solid white;
-  }
-
-  .horizontal {
-    flex-direction: column;
-    .header-primary {
-      padding-top: 1rem;
+    width: 32%;
+    &:first-child {
+      width: 32%;
     }
-
-    .card__img-container {
-      width: 100%;
-      height: 300px;
+    &:nth-child(7n) {
+      width: 32%;
     }
-
-    .card__text-box {
-      width: 100%;
-      padding-left: 0rem;
+    &:nth-child(6n) {
+      width: 32%;
+      flex-direction: column;
+      .card__img-container {
+        width: 100%;
+        height: 300px;
+      }
+      .card__text-box {
+        width: 100%;
+      }
     }
   }
 }
