--- Part 1/client/src/components/grid/Grid.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import './Grid.scss';
-
-const Grid = props => {
-  return <div className="grid">{props.children}</div>;
-};
-
-export default Grid;
