--- part 2/client/src/components/grid/Grid.scss
@@ -1,3 +1,0 @@
-.grid {
-  display: flex;
-}
